Preserve fractional pointer positions across sketch input

+++ lib/prismel/frame.ml
@@ -10,8 +10,8 @@ type t = {
   dt : float;
   fps : float;
   count : int;
-  mouse : int * int;
-  mouse_delta : int * int;
+  mouse : float * float;
+  mouse_delta : float * float;
   keys : Input.key list;
   mouse_buttons : Input.mouse_button list;
   events : Event.t list;
@@ -20,4 +20,3 @@ type t = {
 let key_down key frame = List.mem key frame.keys
 let mouse_down button frame = List.mem button frame.mouse_buttons
 let has_event predicate frame = List.exists predicate frame.events
-
